Finished Problem 123.

+++ Problem0123.h
@@ -11,7 +11,7 @@ using namespace std;
 
 class Problem0123 {
 public:
-    int maxProfit(const vector<int> &prices) {
+    int maxProfit(const vector<int> &prices) {  // 前后缀分解：预处理前缀，枚举后缀
         int f[prices.size()];
         int preMin = prices[0];
         // 1. 预处理f[i]：在第0～i天进行一次交易的最大获利
@@ -25,7 +25,7 @@ public:
         int maxNext = prices[prices.size() - 1];
         // 2. 反向扫描，寻找在第i天买入（一次交易）的最大获利
         result = max(result, f[prices.size() - 1]); // 边界情况: 不进行第2次交易
-        for (int i = prices.size() - 1; i >= 1; --i) {
+        for (int i = (int) prices.size() - 1; i >= 1; --i) {
             result = max(result, f[i - 1] + maxNext - prices[i]);
             maxNext = max(maxNext, prices[i]);
         }
